Split parser function forms
